AEA-3765 Automation Reports are not showing when tests have failed (#36)
## Summary

- Routine Change

The Regression Tests workflow is not generating a report after a test
case fails. Ideally, we would like to be able to trace run which have
failed.

Steps to reproduce the error

Open the [Regression
Tests](https://github.com/NHSDigital/electronic-prescription-service-api-regression-tests/actions/workflows/regression_tests.yml)
GitHub Actions page.
Run the workflow using run-me-for-failed-test branch (not main).
Check the [Reports web
page](https://ubiquitous-adventure-p8885yq.pages.github.io/) to see the
new report.

Goal
The new report is generated showing the failed tests.
There should also be a mention on the graph with regression test runs
which have failed in the past. See screenshot graph below.

### Details

Set default environment to run against to DEV instead of INT.
Change to workflow to allow it to continue to run even if the regression
tests step fails but not any other.

## Reviews Required

- [x] Dev
- [ ] Test
- [ ] Tech Author
- [ ] Product Owner

## Review Checklist

:information_source: This section is to be filled in by the
**reviewer**.

- [ ] I have reviewed the changes in this PR and they fill all or part
of the acceptance criteria of the ticket, and the code is in a mergeable
state.
- [ ] If there were infrastructure, operational, or build changes, I
have made sure there is sufficient evidence that the changes will work.
- [ ] I have ensured the jira ticket has been updated with the github
pull request link