AEA-3822 create ping bdd test for apigee pfp code (#74)
## Summary

**Remove items from this list if they are not relevant. Remove this line
once this has been done**

- :robot: Operational or Infrastructure Change
- :sparkles: New Feature
- :warning: Potential issues that might be caused by this change

### Details

updated structure to support additional products in features. product is
specified using a --product argument in the command line, this then
converts into a tag for the behave command. only features with this tag
will then be run. Any others will be considered skipped. Skipped tests
are hidden to prevent confusion