New: [AEA-4802] - Site org information cards (#291)
## Summary

- Routine Change

### Details

Basic tests. Note that this is only creating the prescription details
page so that the cards have somewhere to live - the actual surrounding
page is way beyond the scope of this ticket, so ONLY the cards are
checked for.

Also, since we have no live data yet, I've added a dev link to take us
to the prescription details page. This will need to be removed later,
and I've left a fixme note to that effect.

---------

Co-authored-by: Sean Steberis <103416906+seansteberisal@users.noreply.github.com>